fix possibility of integer overflow since numberOfItemsToExport from RangeFactory is a long type

+++ neptune-export/src/main/java/com/amazonaws/services/neptune/io/Status.java
@@ -12,18 +12,18 @@ permissions and limitations under the License.
 
 package com.amazonaws.services.neptune.io;
 
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.Supplier;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Supplier;
 
 public class Status {
 
     private static final Logger logger = LoggerFactory.getLogger(Status.class);
 
-    private final AtomicInteger counter = new AtomicInteger();
+    private final AtomicLong counter = new AtomicLong();
     private final AtomicBoolean allowContinue = new AtomicBoolean(true);
     private final StatusOutputFormat outputFormat;
     private final String description;
@@ -44,7 +44,7 @@ public class Status {
     }
 
     public void update() {
-        int counterValue = counter.incrementAndGet();
+        long counterValue = counter.incrementAndGet();
         if (counterValue % 10000 == 0 && outputFormat == StatusOutputFormat.Dot) {
             System.err.print(".");
         } else if (counterValue % 100000 == 0 && outputFormat == StatusOutputFormat.Description) {
